+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.0.3
+- Update `google_generative_ai` dependency
+- Remove workaround for http client override issue in `google_generative_ai`
+  ([#64](https://github.com/google/generative-ai-dart/issues/64))
 - Update README.md
 
 ## 0.0.2