+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Update `google_generative_ai` dependency
 - Remove workaround for http client override issue in `google_generative_ai`
   ([#64](https://github.com/google/generative-ai-dart/issues/64))
+- Add more information to Gemini API not available in user region error message
 - Update README.md
 
 ## 0.0.2